moving utils to dem folder
